style 전체 코드 formattng 및 validator, generator 패키징
