+++ src/main/java/ladder/AppController.java
@@ -1,6 +1,7 @@
 package ladder;
 
 import ladder.model.*;
+import ladder.model.generator.MemberGenerator;
 import ladder.view.InputView;
 import ladder.view.OutputView;
 
@@ -17,7 +18,7 @@ public class AppController {
 
     private LadderGame initGame() {
         String[] names = InputView.inputNames();
-        List<Member> members = Members.generateMembers(names);
+        List<Member> members = MemberGenerator.generateMembers(names);
         List<String> results = Arrays.asList(InputView.inputResults(members.size()));
         int ladderHeight = InputView.inputLadderHeight();
 
@@ -26,7 +27,7 @@ public class AppController {
 
     private EndResult playGame(LadderGame ladderGame){
         OutputView.printLadder(ladderGame);
-        return ladderGame.excuteGame();
+        return ladderGame.executeGame();
     }
 
     private void endGame(EndResult results) {
