+++ src/main/java/ladder/model/generator/MemberGenerator.java
@@ -1,9 +1,18 @@
-package ladder.model;
+package ladder.model.generator;
+
+import ladder.model.Member;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class Members {
+public class MemberGenerator {
+
+    private static final String COMMA = ",";
+
+    public static String[] refineNames(String inputText) {
+        String[] members = inputText.split(COMMA);
+        return members;
+    }
 
     public static List<Member> generateMembers(String[] names) {
         List<Member> members = new ArrayList<>();
