+++ src/main/java/ladder/model/validator/LadderValidator.java
@@ -1,4 +1,4 @@
-package ladder.model;
+package ladder.model.validator;
 
 public class LadderValidator {
     private static final int MIN_HEIGHT = 1;
