+++ src/test/java/ladder/model/MemberValidatorTest.java
@@ -1,5 +1,6 @@
 package ladder.model;
 
+import ladder.model.validator.MemberValidator;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
